WAC/webauthguard/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
httpx==0.27.2
rdflib==7.0.0
opensearch-py==2.6.0
requests==2.32.3
PyJWT==2.9.0
python-multipart==0.0.9